+++ plugins/webapi/autopeering/plugin.go
@@ -18,7 +18,7 @@ import (
 )
 
 // PluginName is the name of the web API autopeering endpoint plugin.
-const PluginName = "WebAPI autopeering Endpoint"
+const PluginName = "WebAPIAutopeeringEndpoint"
 
 var (
 	// plugin is the plugin instance of the web API autopeering endpoint plugin.